Restructure Examples to make it more readable
