+++ src/routes/components/table/examples/TableGroups.svelte
@@ -1,0 +1,24 @@
+<script lang="ts">
+	import CodeContainer from '$docs/components/CodeContainer.svelte';
+	import Table from '$lib/components/Table/Table.svelte';
+	import { groupHTML, groupStoreCode } from '../data/codeBlocks';
+	import { groupsStore } from '../data/data';
+	import type { Group } from '../data/data';
+	import type { TableConfig } from '$lib/models/Models';
+
+	const groupConfig: TableConfig<Group> = {
+		id: 'userGroups',
+		data: groupsStore,
+		columns: {
+			id: {
+				header: 'ID'
+			}
+		}
+	};
+</script>
+
+<div id="groups">
+	<CodeContainer title="Table component" svelte={groupHTML} data={groupStoreCode}>
+		<Table config={groupConfig} />
+	</CodeContainer>
+</div>
